REFACTOR: constants 및 utils 모듈 분리
