+++ src/CarRacingManager.js
@@ -1,7 +1,5 @@
-function sleeping(ms) {
-  const wakeUpTime = Date.now() + ms;
-  while (Date.now() < wakeUpTime) {}
-}
+import { ERROR_MESSAGES, GAME_MESSAGES } from "./constants";
+import { sleeping } from "./utils";
 
 export class CarRacingManager {
   #canMoveCondition = 4;
@@ -11,11 +9,11 @@ export class CarRacingManager {
 
   constructor() {}
 
-  gameStart(names, endProcess, wait) {
+  gameStart(names, endProcess, wait = sleeping) {
     try {
       this.#setNamesAndMoves(names);
       console.log("\n실행결과");
-      this.roundInterval(endProcess, wait || sleeping);
+      this.roundInterval(endProcess, wait);
     } catch (error) {
       this.gameEnd(endProcess, error.message);
     }
@@ -104,18 +102,3 @@ export class CarRacingManager {
     return Math.floor(Math.random() * 10);
   }
 }
-
-export const ERROR_MESSAGES = {
-  OVER_MAXIMUM_NAME_LENGTH: "이름은 5자 이하만 가능합니다.",
-  WINNERS_TYPE_IS_NOT_ARRAY: "winners type is not Array",
-};
-
-export const GAME_MESSAGES = {
-  QUESTION:
-    "경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).\n",
-  GAME_OVER: "게임이 종료되었습니다.",
-};
-
-export const INTEGERS_UNDER_TEN = Array(10)
-  .fill(0)
-  .map((v, i) => v + i);
