+++ src/utils.js
@@ -1,0 +1,4 @@
+export const sleeping = (ms) => {
+  const wakeUpTime = Date.now() + ms;
+  while (Date.now() < wakeUpTime) {}
+};
